Fix Bugs
亟待解决的问题：
th：if标签在a标签中失效

+++ src/main/java/com/iotdreamclub/demo/controller/BillController.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.extension.api.R;
 import com.iotdreamclub.demo.dao.MatchBillMapper;
 import com.iotdreamclub.demo.service.BillService;
 import com.iotdreamclub.demo.service.MatchBillService;
+import org.apache.shiro.session.Session;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,7 +12,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.sql.Date;
 
 //资金管理相关操作
@@ -59,8 +62,23 @@ public class BillController {
 
     @RequestMapping("deleteBillInfo/{billId}")
     @ResponseBody
-    public String deleteBillInfo(@PathVariable Long billId , HttpServletResponse response){
-        billService.deleteBillInfo(billId);
+    public String deleteBillInfo(@PathVariable Long billId , HttpServletResponse response , HttpSession session) throws IOException {
+
+        String userLimit = String.valueOf(session.getAttribute("limit"));
+        System.out.println(userLimit);
+
+        if(String.valueOf(session.getAttribute("limit")).equals("2") || String.valueOf(session.getAttribute("limit")).equals("3")){
+            response.setContentType("text/html; charset=UTF-8"); //转码
+            PrintWriter out = response.getWriter();
+            out.flush();
+            out.println("<script>");
+            out.println("alert('权限不足无法操作');");
+            out.println("history.back();");
+            out.println("</script>");
+            response.sendRedirect("/economics_management");
+        }else {
+            billService.deleteBillInfo(billId);
+        }
         try {
             response.sendRedirect("/economics_management");
         }catch (Exception e){
